FingerTree: move some Require at beginning of a file, which solves an Obligation Tactic issue
After simplifiying dependencies in the trunk, Program.Tactics was not loaded
in DependentFingerTreeModule.v until the Require Omega in the middle of the file,
and doing that was setting Obligation Tactic back to program_simpl instead of
local splitTac.

Anyway, a Require in the middle of an internal module is a _bad_ habit.



git-svn-id: svn://scm.gforge.inria.fr/svnroot/coq-contribs/trunk@1147 0cf17b13-060f-0410-b1b1-c666bec9822a